# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Swagger API.",
        "title": "Swagger Demo",
        "contact": {
            "name": "菜鳥工程師肉豬",
            "url": "https://matthung0807.blogspot.com/"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "paths": {
        "/employee": {
            "post": {
                "description": "Create a new employee",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "employee"
                ],
                "summary": "Create a new employee",
                "parameters": [
                    {
                        "description": "create params",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/view.CreateEmployeeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/view.CreateEmployeeResponse"
                        }
                    }
                }
            }
        },
        "/hello": {
            "get": {
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "Hello"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "user name",
                        "name": "name",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "view.CreateEmployeeRequest": {
            "type": "object",
            "properties": {
                "age": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "title": {
                    "description": "Employee title:\n* engineer - developer, programer.\n* manager - product manger, project manager, supervisor.",
                    "type": "string",
                    "enum": [
                        "engineer",
                        "manager"
                    ]
                }
            }
        },
        "view.CreateEmployeeResponse": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                }
            }
        }
    }
}